requirements.txt
# Core data processing libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Machine learning libraries
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.0

# Deep learning (if needed)
tensorflow==2.13.0
torch==2.0.1
transformers==4.31.0

# NLP and text processing
nltk==3.8.1
spacy==3.6.1
gensim==4.3.1

# Feature engineering
featuretools==1.26.0
feature-engine==1.6.1
category_encoders==2.6.1

# AutoML
optuna==3.3.0
hyperopt==0.2.7
ray[tune]==2.6.3

# Explainability
shap==0.42.1
lime==0.2.0.1
eli5==0.13.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# LLM integration
openai==0.28.0
langchain==0.0.267
sentence-transformers==2.2.2

# Utilities
tqdm==4.66.1
joblib==1.3.2
dill==0.3.7
pyyaml==6.0.1

# GPU acceleration
cupy-cuda11x==12.2.0  # Only if CUDA is available 